requirements.txt
pillow
PyMuPDF
inquirer